requirements.txt
# Core Django (Python 3.13 compatible versions)
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0

# Database
psycopg2-binary>=2.9.0,<3.0.0
dj-database-url>=2.1.0,<3.0.0

# CORS
django-cors-headers>=4.3.0,<5.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# File handling (Pillow will be installed separately)
# Pillow>=10.3.0
django-cleanup>=8.0.0,<9.0.0

# PDF processing for plagiarism check and UDK certificate
PyPDF2>=3.0.0,<4.0.0
pdfplumber>=0.10.0,<1.0.0
reportlab>=4.0.0,<5.0.0
qrcode[pil]>=7.0,<8.0

# Word documents (UDK standalone: doc, docx)
docx2txt>=0.8,<1.0.0

# API integrations
requests>=2.31.0,<3.0.0
google-generativeai>=0.3.0,<1.0.0

# Validation
django-phonenumber-field>=7.3.0,<8.0.0
phonenumbers>=8.13.0,<9.0.0

# Utilities
celery>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0
django-celery-beat>=2.5.0,<3.0.0

# Telegram Bot
python-telegram-bot>=20.7,<21.0.0

# Production
gunicorn>=21.2.0,<22.0.0
whitenoise>=6.6.0,<7.0.0

# Development (optional)
# django-debug-toolbar>=4.2.0,<5.0.0